fix(timestamp): pseudo-fix for timestamps being CET in stead of UTC
The timestamp we get from teracom is CET, which has changes during the year due to daylight savings.
For this reason we should convert to UTC. For simplicity this method only subtracts an hour, which will be accurate
until daylight savings ends in the spring, after which it will be off by one.
For more details, see the comment in GatewayService in this commit.